Refactor viewer2d PDF exporter into module skeleton
