+++ viewer2d/pdf/pdf_writer.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <filesystem>
+#include <string>
+#include <vector>
+
+struct PdfObject {
+  std::string body;
+};
+
+bool WritePdfDocument(const std::filesystem::path &outputPath,
+                      const std::vector<PdfObject> &objects,
+                      size_t catalogObjectIndex, std::string &error);
